inference_model/requirements.txt
torch==2.5.1
torchvision==0.20.1
pillow==11.0.0
tritonclient[grpc]==2.51.0